refactor: 不要な型定義とフックを削除し、編集モードアクションを新規作成
